Refactor event and course details: extract components, unify admin toolbar, improve layout
- Split EventDetails into banner, info grid, ticket form, my-teams section
- Extract CourseDetails enrolled panel into EnrolledPanel component
- Lift MyTeamRow to shared component with readOnly prop; Profile owns team editing, event page shows read-only summary with 'Manage in Profile' CTA
- Add CourseHeader/CourseRoot mirroring EventHeader/EventRoot; toolbars now include Edit on detail pages alongside Create
- Mobile-first order swap: CTA column above fold on small screens
- Reserve image aspect-ratio to prevent CLS, standardize loading/error states, add prominent Facebook share button on both detail pages
