+++ src/pages/payments/CancelPage.jsx
@@ -31,7 +31,21 @@ export default function CancelPage() {
         <p className="text-base-content mb-8">
           Your transaction has been cancelled. If this was a mistake, you can try again.
         </p>
-        <Button variant="primary" onClick={handleGoBack}>
+        <Button variant="secondary" onClick={handleGoBack}>
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            className="h-5 w-5 mr-2"
+            fill="none"
+            viewBox="0 0 24 24"
+            stroke="currentColor"
+          >
+            <path
+              strokeLinecap="round"
+              strokeLinejoin="round"
+              strokeWidth={2}
+              d="M10 19l-7-7m0 0l7-7m-7 7h18"
+            />
+          </svg>
           Go Back to Home
         </Button>
       </GlassCard>
