+++ src/pages/tickets/TicketPage.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from "react";
-import { Link, useParams, useNavigate, useSearchParams } from "react-router-dom";
-import { isAuthenticated, getAuthToken } from "../../auth/auth";
+import { useParams, useNavigate, useSearchParams } from "react-router-dom";
+import { isAuthenticated, fetchWithAuth } from "../../auth/auth";
 import TicketCard from "../../components/tickets/TicketCard";
-import { PageContainer, GlassCard, Spinner } from "../../components/ui";
+import { Button, PageContainer, GlassCard, Spinner } from "../../components/ui";
 
 export default function TicketPage() {
   const { ticketCode } = useParams();
@@ -21,12 +21,9 @@ export default function TicketPage() {
       navigate("/login");
       return;
     }
-    const token = getAuthToken();
     Promise.all(
       codes.map((code) =>
-        fetch(`${import.meta.env.VITE_DEV_URI}tickets/${code}`, {
-          headers: { Authorization: "Bearer " + token },
-        }).then(async (res) => {
+        fetchWithAuth(`${import.meta.env.VITE_DEV_URI}tickets/${code}`).then(async (res) => {
           if (!res.ok) throw new Error(`Ticket ${code} not found`);
           return res.json();
         })
@@ -61,9 +58,23 @@ export default function TicketPage() {
       <PageContainer center>
         <GlassCard className="p-8 text-center max-w-sm border border-red-100">
           <p className="text-red-500 font-medium mb-4">{error}</p>
-          <Link to="/profile" className="text-sm text-base-content/70 hover:underline">
+          <Button variant="secondary" to="/profile">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              className="h-5 w-5 mr-2"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth={2}
+                d="M10 19l-7-7m0 0l7-7m-7 7h18"
+              />
+            </svg>
             Back to profile
-          </Link>
+          </Button>
         </GlassCard>
       </PageContainer>
     );
@@ -73,20 +84,25 @@ export default function TicketPage() {
   return (
     <PageContainer>
       <div className="max-w-lg mx-auto py-10 px-4">
-        <Link
-          to="/profile"
-          className="inline-flex items-center gap-1.5 text-sm text-base-content/50 hover:text-base-content/70 mb-6 transition-colors print:hidden"
-        >
-          <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-            <path
-              strokeLinecap="round"
-              strokeLinejoin="round"
-              strokeWidth={2}
-              d="M15 19l-7-7 7-7"
-            />
-          </svg>
-          Back to profile
-        </Link>
+        <div className="mb-6 print:hidden">
+          <Button variant="secondary" to="/profile">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              className="h-5 w-5 mr-2"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth={2}
+                d="M10 19l-7-7m0 0l7-7m-7 7h18"
+              />
+            </svg>
+            Back to profile
+          </Button>
+        </div>
         {isMulti && (
           <p className="text-xs text-base-content/50 mb-4 print:hidden">
             Showing {tickets.length} tickets from this order
